+++ src/name-uniqueness-check.ts
@@ -4,16 +4,19 @@ import { ValidationErrorCodes } from "./error-codes";
 /**
  * Checks if all provided Elements have a unique property
  */
-export const propertyUniquenessCheck =
- <T extends object>(namedElementList : T[], propertyName : keyof T, entityName : string) => {
+export const propertyUniquenessCheck = <T extends object>(
+  namedElementList : T[],
+  propertyName : keyof T,
+  entityName : string,
+) : void => {
   const nameList = [];
 
   namedElementList.forEach((namedElement) => {
     if (nameList.includes(namedElement[propertyName])) {
-      throw Error(error(ValidationErrorCodes.V27) +
+      throw Error(error(ValidationErrorCodes.failedNameUniquenessCheck) +
         `- "${entityName}" - ${highlight(`"${namedElement[propertyName]}"`)}`);
     }
 
     nameList.push(namedElement[propertyName]);
-  })
-}
+  });
+};