+++ src/validate-package-configuration.ts
@@ -1,0 +1,19 @@
+import { success } from "./chalk-formatting";
+import { isMetaPackage } from "./is-meta-package";
+import { ValidateMetaPackageFunctionDefinition } from "./validate-meta-package-function-definition";
+import { buildAllFunctionDefinitions } from "./build-all-functions-definitions";
+
+/** Validates the content of a `meta-package.json` file */
+export const validatePackageConfiguration = async (configurationData : unknown) : Promise<void> => {
+  isMetaPackage(configurationData);
+
+  const builtPackage = {
+    ...configurationData,
+    functionsDefinitions: await buildAllFunctionDefinitions(configurationData.functionsDefinitions),
+  };
+
+  await new ValidateMetaPackageFunctionDefinition(builtPackage)
+    .execute();
+
+  console.log(success("Final Package file passed validation."));
+};