+++ test/test-files/example-class.js
@@ -1,0 +1,12 @@
+class MainClass {
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  constructor (validatorParam) {
+    if (validatorParam !== "n19oenvzc9vc0ui4n5109b8n1") {
+      throw Error("Class not instantiated with the expected parameter");
+    }
+  }
+}
+
+module.exports = {
+  MainClass,
+};